BE/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
openai==0.28.1
python-jose==3.3.0
cryptography==42.0.2 